// src/libs/openapi/models/DialogueUserMessageDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DialogueUserMessageDto = {
    appId?: string;
    sessionId?: string;
    gender?: string;
    language?: string;
    llm?: string;
    message: string;
};